requirements.txt
altgraph==0.17.4
asn1crypto==1.5.1
bip-utils==2.9.3
cbor2==5.6.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coincurve==20.0.0
crcmod==1.7
ecdsa==0.19.0
ed25519-blake2b==1.4.1
idna==3.10
macholib==1.16.3
more-itertools==10.5.0
packaging==24.2
pillow==11.0.0
py-sr25519-bindings==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyNaCl==1.5.0
requests==2.32.3
six==1.17.0
urllib3==2.2.3
